# src/main/java/com/customermanagement/dto/CustomerRequest.java
package com.customermanagement.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotNull;
import lombok.Data;
import java.time.LocalDateTime;

@Data 
public class CustomerRequest {
    @NotNull(message = "Name is required")
    private String name;

    @NotNull(message = "Email is required")
    @Email(message = "Email should be valid")
    private String email;

    private Double annualSpend;

    private LocalDateTime lastPurchaseDate;
}